requirements.txt
streamlit>=1.24
langchain>=0.2
langchain-community>=0.0.###  # optional: the new community package if you use it
faiss-cpu
sentence-transformers
transformers
torch
duckduckgo_search
python-dotenv
groq
